LPS-88361 Merge FindActionHelper into FindStrutsAction
