--- modules/apps/message-boards/message-boards-web/src/main/java/com/liferay/message/boards/web/internal/portlet/action/CategoryFindActionHelper.java
@@ -1,93 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.message.boards.web.internal.portlet.action;
-
-import com.liferay.message.boards.model.MBCategory;
-import com.liferay.message.boards.service.MBCategoryLocalService;
-import com.liferay.portal.kernel.portlet.PortletLayoutFinder;
-import com.liferay.portal.struts.BaseFindActionHelper;
-import com.liferay.portal.struts.FindActionHelper;
-
-import javax.portlet.PortletURL;
-
-import javax.servlet.http.HttpServletRequest;
-
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
-
-/**
- * @author Adolfo Pérez
- */
-@Component(
-	immediate = true,
-	property = "model.class.name=com.liferay.message.boards.model.MBCategory",
-	service = FindActionHelper.class
-)
-public class CategoryFindActionHelper extends BaseFindActionHelper {
-
-	@Override
-	public long getGroupId(long primaryKey) throws Exception {
-		MBCategory category = _mbCategoryLocalService.getCategory(primaryKey);
-
-		return category.getGroupId();
-	}
-
-	@Override
-	public String getPrimaryKeyParameterName() {
-		return "mbCategoryId";
-	}
-
-	@Override
-	public void setPrimaryKeyParameter(PortletURL portletURL, long primaryKey)
-		throws Exception {
-
-		portletURL.setParameter(
-			getPrimaryKeyParameterName(), String.valueOf(primaryKey));
-	}
-
-	@Override
-	protected void addRequiredParameters(
-		HttpServletRequest request, String portletId, PortletURL portletURL) {
-
-		portletURL.setParameter(
-			"mvcRenderCommandName", "/message_boards/view_category");
-	}
-
-	@Override
-	protected PortletLayoutFinder getPortletLayoutFinder() {
-		return _portletPageFinder;
-	}
-
-	@Reference(unbind = "-")
-	protected void setMBCategoryLocalService(
-		MBCategoryLocalService mbCategoryLocalService) {
-
-		_mbCategoryLocalService = mbCategoryLocalService;
-	}
-
-	@Reference(
-		target = "(model.class.name=com.liferay.message.boards.model.MBCategory)",
-		unbind = "-"
-	)
-	protected void setPortletLayoutFinder(
-		PortletLayoutFinder portletPageFinder) {
-
-		_portletPageFinder = portletPageFinder;
-	}
-
-	private MBCategoryLocalService _mbCategoryLocalService;
-	private PortletLayoutFinder _portletPageFinder;
-
-}
