+++ modules/apps/message-boards/message-boards-web/src/main/java/com/liferay/message/boards/web/internal/portlet/action/FindMessageAction.java
@@ -15,11 +15,15 @@
 package com.liferay.message.boards.web.internal.portlet.action;
 
 import com.liferay.message.boards.constants.MBPortletKeys;
+import com.liferay.message.boards.model.MBMessage;
+import com.liferay.message.boards.service.MBMessageLocalService;
+import com.liferay.portal.kernel.portlet.PortletLayoutFinder;
 import com.liferay.portal.kernel.struts.StrutsAction;
-import com.liferay.portal.struts.FindActionHelper;
+import com.liferay.portal.struts.FindStrutsAction;
+
+import javax.portlet.PortletURL;
 
 import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -35,26 +39,59 @@ import org.osgi.service.component.annotations.Reference;
 	},
 	service = StrutsAction.class
 )
-public class FindMessageAction implements StrutsAction {
+public class FindMessageAction extends FindStrutsAction {
 
 	@Override
-	public String execute(
-			HttpServletRequest request, HttpServletResponse response)
+	public long getGroupId(long primaryKey) throws Exception {
+		MBMessage message = _mbMessageLocalService.getMessage(primaryKey);
+
+		return message.getGroupId();
+	}
+
+	@Override
+	public String getPrimaryKeyParameterName() {
+		return "messageId";
+	}
+
+	@Override
+	public void setPrimaryKeyParameter(PortletURL portletURL, long primaryKey)
 		throws Exception {
 
-		_findActionHelper.execute(request, response);
+		portletURL.setParameter(
+			getPrimaryKeyParameterName(), String.valueOf(primaryKey));
+	}
 
-		return null;
+	@Override
+	protected void addRequiredParameters(
+		HttpServletRequest request, String portletId, PortletURL portletURL) {
+
+		portletURL.setParameter(
+			"mvcRenderCommandName", "/message_boards/view_message");
+	}
+
+	@Override
+	protected PortletLayoutFinder getPortletLayoutFinder() {
+		return _portletPageFinder;
+	}
+
+	@Reference(unbind = "-")
+	protected void setMBMessageLocalService(
+		MBMessageLocalService mbMessageLocalService) {
+
+		_mbMessageLocalService = mbMessageLocalService;
 	}
 
 	@Reference(
 		target = "(model.class.name=com.liferay.message.boards.model.MBMessage)",
 		unbind = "-"
 	)
-	protected void setFindActionHelper(FindActionHelper findActionHelper) {
-		_findActionHelper = findActionHelper;
+	protected void setPortletLayoutFinder(
+		PortletLayoutFinder portletPageFinder) {
+
+		_portletPageFinder = portletPageFinder;
 	}
 
-	private FindActionHelper _findActionHelper;
+	private MBMessageLocalService _mbMessageLocalService;
+	private PortletLayoutFinder _portletPageFinder;
 
 }
